evaluation/requirements.txt
sentence-transformers
faiss-cpu
numpy
scipy
transformers
datasets
openai
tqdm
python-dotenv
